data-manager: Migrate from Play framework to Dropwizard (#133, #83)
It consists mostly of:

- Migration from sbt to maven. This includes:
  - Support for running tests using scalatest
  - Support for incremental and continuos compilation and testing
  - Proper merging of akka configuration into fat jar
- Translation of controllers
- Translation of integration tests
- Simplifying dependences (for example in the case of logging)
- Changing data serialization from (json -> binary) to (json -> text)
  This will allow any application to restore the data if necessary

Still pending:

- Migration of the configuration system
- Proper dependency injection
- Document key steps of the migration
