+++ benchflow-data-manager/.gitignore
@@ -2,6 +2,8 @@
 .settings
 .classpath
 .project
+.cache-main
+.cache-tests
 
 #Idea stuff
 .idea
@@ -17,6 +19,8 @@ application/src/test/resources/configuration.properties
 
 benchflow_make
 /bin/
+
+StoredCredential
 
 #sbt/play
 logs
