--- benchflow-data-manager/conf/application.conf
@@ -1,86 +1,0 @@
-# This is the main configuration file for the application.
-# ~~~~~
-
-# Secret key
-# ~~~~~
-# The secret key is used to secure cryptographics functions.
-# If you deploy your application to several instances be sure to use the same key!
-play.crypto.secret="changeme"
-play.crypto.secret=${?APPLICATION_SECRET}
-
-# The application languages
-# ~~~~~
-play.i18n.langs=["en"]
-
-# Global object class
-# ~~~~~
-# Define the Global object class for this application.
-# Defaults to Global in the root package.
-# application.global=my.Global
-
-# Router
-# ~~~~~
-# Define the Router object to use for this application.
-# This router will be looked up first when the application is starting up,
-# so make sure this is the entry point.
-# Furthermore, it's assumed your route file is named properly.
-# So for an application router like `my.application.Router`,
-# you may need to define a router file `conf/my.application.routes`.
-# Default to Routes in the root package (and conf/routes)
-# application.router=my.application.Routes
-
-# Database configuration
-# ~~~~~
-# You can declare as many datasources as you want.
-# By convention, the default datasource is named `default`
-#
-# db.default.driver=org.h2.Driver
-# db.default.url="jdbc:h2:mem:play"
-# db.default.user=sa
-# db.default.password=""
-
-# Evolutions
-# ~~~~~
-# You can disable evolutions if needed
-# evolutionplugin=disabled
-
-# Uncomment this for the most verbose Akka debugging:
-#akka {
-#    loglevel = "DEBUG"
-#    actor {
-#        debug {
-#            receive = on
-#            autoreceive = on
-#            lifecycle = on
-#        }
-#    }
-#}
-
-# Google Drive configuration
-googledrive {
-  appname = ${?GOOGLE_DRIVE_APPNAME}
-  secret = ${?GOOGLE_DRIVE_SECRET}
-  credentials_dir = "./"
-  credentials_dir = ${?GOOGLE_DRIVE_CREDENTIALS_DIR}
-  baseFolder = "benchflow"
-  baseFolder = ${?GOOGLE_DRIVE_BASEFOLDER}
-}
-
-cassandra {
-  host = ["localhost"]
-  host = ${?CASSANDRA_HOST}
-  keyspace = "benchflow"
-  keyspace = ${?CASSANDRA_KEYSPACE}
-  username = ""
-  username = ${?CASSANDRA_USERNAME}
-  password = ""
-  password = ${?CASSANDRA_PASSWORD}
-}
-
-minio {
-  url = ${?MINIO_URL}
-  accessKey = ${?MINIO_ACCESS_KEY}
-  secretKey = ${?MINIO_SECRET_KEY}
-  defaultBucket = "test"
-  defaultBucket = ${?MINIO_DEFAULT_BUCKET}
-}
