--- benchflow-data-manager/conf/prod.conf
@@ -1,3 +1,0 @@
-include "application.conf"
-
-play.server.http.port = 8080
