--- benchflow-data-manager/project/build.properties
@@ -1,4 +1,0 @@
-#Activator-generated Properties
-#Sun May 22 16:40:23 CEST 2016
-template.uuid=bef7dbe2-90ea-415e-89e1-c1ba7318a455
-sbt.version=0.13.8
