--- benchflow-data-manager/project/plugins.sbt
@@ -1,16 +1,0 @@
-// Comment to get more information during initialization
-logLevel := Level.Warn
-
-// The Typesafe repository
-resolvers += Resolver.typesafeRepo("releases")
-
-// Use the Play sbt plugin for Play projects
-addSbtPlugin("com.typesafe.play" % "sbt-plugin" % "2.5.4")
-
-addSbtPlugin("com.typesafe.sbt" % "sbt-less" % "1.1.0")
-
-addSbtPlugin("com.typesafe.sbt" % "sbt-coffeescript" % "1.0.0")
-
-addSbtPlugin("com.typesafe.sbteclipse" % "sbteclipse-plugin" % "4.0.0")
-
-addSbtPlugin("com.eed3si9n" % "sbt-assembly" % "0.12.0")
