+++ benchflow-data-manager/src/main/resources/application.conf
@@ -1,0 +1,47 @@
+# Uncomment this for the most verbose Akka debugging:
+#akka {
+#    loglevel = "DEBUG"
+#    actor {
+#        debug {
+#            receive = on
+#            autoreceive = on
+#            lifecycle = on
+#        }
+#    }
+#}
+
+# Google Drive configuration
+googledrive {
+  appname = "benchflow-datamanager"
+  appname = ${?GOOGLE_DRIVE_APPNAME}
+  secret = """
+{"installed":{"client_id":"965160099207-k2sj75qnn9kk9mavp0f439hg1shfhdnc.apps.googleusercontent.com","project_id":"innate-plexus-133814","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://accounts.google.com/o/oauth2/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"wezZvVgoyMN6V6nQYyOgLXQc","redirect_uris":["urn:ietf:wg:oauth:2.0:oob","http://localhost"]}}
+"""
+  secret = ${?GOOGLE_DRIVE_SECRET}
+  credentials_dir = "./"
+  credentials_dir = ${?GOOGLE_DRIVE_CREDENTIALS_DIR}
+  baseFolder = "benchflow"
+  baseFolder = ${?GOOGLE_DRIVE_BASEFOLDER}
+}
+
+cassandra {
+  host = ["localhost"]
+  host = ${?CASSANDRA_HOST}
+  keyspace = "benchflow"
+  keyspace = ${?CASSANDRA_KEYSPACE}
+  username = ""
+  username = ${?CASSANDRA_USERNAME}
+  password = ""
+  password = ${?CASSANDRA_PASSWORD}
+}
+
+minio {
+  url = "http://localhost:9001"
+  url = ${?MINIO_URL}
+  accessKey = "YU2J1ONZR1E7MAOCZV1Q"
+  accessKey = ${?MINIO_ACCESS_KEY}
+  secretKey = "TBtf+uoVErl/J3/0tEoICm/3PXfaQE6urWFKTYNI"
+  secretKey = ${?MINIO_SECRET_KEY}
+  defaultBucket = "test"
+  defaultBucket = ${?MINIO_DEFAULT_BUCKET}
+}
